# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Customer 360 Intelligence Platform",
    "description": "\nCustomer 360 Intelligence Platform API.\n\nDeterministic customer data, agent-generated insights and grounded natural-language Q&A over a\nread-only banking dataset. Monetary values are integer cents internally and decimal strings at this\nboundary; every response carries a correlation ID and trace ID in `meta`.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/reports": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Define a report",
        "description": "Create a report definition owned by the caller (task 18.5).",
        "operationId": "define_report_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefineReportBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DefinitionCreated_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{definition_id}/run": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Run a report now",
        "description": "Generate the report now, entitlement re-checked and masked (task 18.2, 18.5).",
        "operationId": "run_report_reports__definition_id__run_post",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Definition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RunView_"
                }
              }
            }
          },
          "404": {
            "description": "No such report, or not owned by the caller"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{definition_id}/runs": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "A report's run history",
        "description": "Every run of a definition the caller owns, newest first (task 18.5).",
        "operationId": "list_report_runs_reports__definition_id__runs_get",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Definition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RunList_"
                }
              }
            }
          },
          "404": {
            "description": "No such report, or not owned by the caller"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/runs/{run_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download a run's artifact",
        "description": "Stream the generated artifact for a run the caller owns (task 18.5).",
        "operationId": "download_run_reports_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/pdf": {}
            }
          },
          "404": {
            "description": "No such run, not owned by the caller, or no artifact"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{definition_id}/schedules": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Attach a schedule to a report",
        "description": "Attach a cron-like cadence to a report the caller owns (task 18.4, 18.5).",
        "operationId": "create_schedule_reports__definition_id__schedules_post",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Definition Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ScheduleCreated_"
                }
              }
            }
          },
          "404": {
            "description": "No such report, or not owned by the caller"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/schedules/{schedule_id}": {
      "delete": {
        "tags": [
          "reports"
        ],
        "summary": "Disable a schedule",
        "description": "Deactivate a schedule so it stops running (task 18.5).",
        "operationId": "disable_schedule_reports_schedules__schedule_id__delete",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_str__bool__"
                }
              }
            }
          },
          "404": {
            "description": "No such schedule"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BrandingBody": {
        "properties": {
          "brand_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Name"
          },
          "tagline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tagline"
          }
        },
        "type": "object",
        "title": "BrandingBody",
        "description": "Per-definition branding (task 18.1). Falls back to platform defaults when omitted."
      },
...
      "DefineReportBody": {
        "properties": {
          "report_type": {
            "$ref": "#/components/schemas/ReportType"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeBody"
          },
          "branding": {
            "$ref": "#/components/schemas/BrandingBody",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "report_type",
          "title",
          "scope"
        ],
        "title": "DefineReportBody",
        "description": "The body of ``POST /reports`` (task 18.5)."
      },
      "DefinitionCreated": {
        "properties": {
          "definition_id": {
            "type": "integer",
            "title": "Definition Id"
          }
        },
        "type": "object",
        "required": [
          "definition_id"
        ],
        "title": "DefinitionCreated"
      },
...
      "Envelope_DefinitionCreated_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DefinitionCreated"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[DefinitionCreated]"
      },
...
      "Envelope_RunList_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunList"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[RunList]"
      },
...
      "Envelope_RunView_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunView"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[RunView]"
      },
      "Envelope_ScheduleCreated_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ScheduleCreated"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[ScheduleCreated]"
      },
...
      "Envelope_dict_str__bool__": {
        "properties": {
          "data": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[dict[str, bool]]"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNAUTHENTICATED",
          "ENTITLEMENT_DENIED",
          "CUSTOMER_NOT_FOUND",
          "VALIDATION_ERROR",
          "AGENT_TIMEOUT",
          "UPSTREAM_UNAVAILABLE",
          "RATE_LIMITED",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "description": "The complete error code set from design \u00a76.3."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Meta": {
        "properties": {
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "masked_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Masked Fields"
          },
          "computed_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Computed Fields"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ModuleError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "correlation_id",
          "as_of"
        ],
        "title": "Meta",
        "description": "Envelope metadata."
      },
      "ModuleError": {
        "properties": {
          "module": {
            "type": "string",
            "title": "Module"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "module",
          "code",
          "message"
        ],
        "title": "ModuleError",
        "description": "A single domain module's failure, surfaced without failing the whole response."
      },
...
      "ReportScopeKind": {
        "type": "string",
        "enum": [
          "CUSTOMER",
          "BOOK",
          "SEGMENT"
        ],
        "title": "ReportScopeKind",
        "description": "What a report runs over (task 18.1).\n\n* ``CUSTOMER`` \u2014 a single customer, named by ``customer_id``.\n* ``BOOK`` \u2014 an explicit set of customer ids the owner is entitled to.\n* ``SEGMENT`` \u2014 every customer in one or more segments, within the owner's entitlement."
      },
      "ReportType": {
        "type": "string",
        "enum": [
          "PDF_PACK",
          "MEETING_BRIEFING"
        ],
        "title": "ReportType",
        "description": "The closed set of report kinds Phase 18 produces.\n\n* ``PDF_PACK`` \u2014 a branded 360 PDF pack per customer, assembled from the masked service reads\n  (task 18.2).\n* ``MEETING_BRIEFING`` \u2014 a prepare-for-meeting briefing composing the 360 summary, open signals\n  and cited talking points (task 18.3)."
      },
      "RunList": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunView"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "runs"
        ],
        "title": "RunList"
      },
...
      "RunView": {
        "properties": {
          "run_id": {
            "type": "integer",
            "title": "Run Id"
          },
          "definition_id": {
            "type": "integer",
            "title": "Definition Id"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trigger_kind": {
            "type": "string",
            "title": "Trigger Kind"
          },
          "customers_rendered": {
            "type": "integer",
            "title": "Customers Rendered"
          },
          "model_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Model Ids"
          },
          "prompt_versions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prompt Versions"
          },
          "degraded": {
            "type": "boolean",
            "title": "Degraded"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "definition_id",
          "report_type",
          "status",
          "trigger_kind",
          "customers_rendered",
          "model_ids",
          "prompt_versions",
          "degraded",
          "started_at",
          "finished_at",
          "error"
        ],
        "title": "RunView",
        "description": "A run's status and provenance, value-free (task 18.5)."
      },
      "ScheduleBody": {
        "properties": {
          "cadence": {
            "type": "string",
            "pattern": "^(DAILY|WEEKLY|MONTHLY)$",
            "title": "Cadence"
          }
        },
        "type": "object",
        "required": [
          "cadence"
        ],
        "title": "ScheduleBody",
        "description": "The body of ``POST /reports/{id}/schedules`` (task 18.5)."
      },
      "ScheduleCreated": {
        "properties": {
          "schedule_id": {
            "type": "integer",
            "title": "Schedule Id"
          },
          "definition_id": {
            "type": "integer",
            "title": "Definition Id"
          },
          "cadence": {
            "type": "string",
            "title": "Cadence"
          }
        },
        "type": "object",
        "required": [
          "schedule_id",
          "definition_id",
          "cadence"
        ],
        "title": "ScheduleCreated"
      },
      "ScopeBody": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ReportScopeKind"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "customer_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Customer Ids",
            "default": []
          },
          "segments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Segments",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "ScopeBody",
        "description": "The scope a report runs over (task 18.1)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}